+++ src/view/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useYfmShowElemWithId';
+export * from './useYfmHtmlBlockStyles';
